compositeerrors: move package to services/ctl-api/internal/pkg/compositeerrors
Amp-Thread-ID: https://ampcode.com/threads/T-019e1b5a-8790-728c-8319-58573ed8cbfa
